user profile design completed
