+++ app/src/main/AndroidManifest.xml
@@ -22,6 +22,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
+        <activity android:name=".CreateProfile"></activity>
         <activity android:name=".PasswordActivity" />
         <activity android:name=".MapActivity" />
