+++ app/src/main/java/com/example/myapplication/RegisterActivity.java
@@ -5,14 +5,12 @@ import androidx.appcompat.app.AppCompatActivity;
 
 import android.content.Context;
 import android.content.Intent;
-import android.media.Image;
 import android.net.ConnectivityManager;
 import android.net.NetworkInfo;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
-import android.widget.ImageButton;
 import android.widget.Toast;
 
 import com.google.android.gms.tasks.OnCompleteListener;
@@ -24,24 +22,24 @@ import com.google.firebase.database.FirebaseDatabase;
 
 public class RegisterActivity extends AppCompatActivity {
 
-    EditText user_name, user_email, user_password, user_confirmPassword, user_mobile;
+    EditText  user_email, user_password, user_confirmPassword;
     Button btn_register;
     FirebaseAuth firebaseAuth;
     FirebaseDatabase firebaseDatabase;
     DatabaseReference reference, reference1;
-    String username, email, password, confirmPwd, mobile;
+    String email, password, confirmPwd;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_register);
 
-        user_name = findViewById(R.id.username);
+
         user_email = findViewById(R.id.email);
         user_password = findViewById(R.id.password);
         user_confirmPassword = findViewById(R.id.etRePassword);
         btn_register = findViewById(R.id.signup);
-        user_mobile = findViewById(R.id.mobile);
+
 
         firebaseAuth = FirebaseAuth.getInstance();
 
@@ -78,27 +76,19 @@ public class RegisterActivity extends AppCompatActivity {
     private Boolean validate(){
         Boolean result = false;
 
-        username = user_name.getText().toString();
+
         email = user_email.getText().toString();
         password = user_password.getText().toString();
         confirmPwd = user_confirmPassword.getText().toString();
-        mobile = user_mobile.getText().toString();
 
-        if(username.isEmpty())
-        {
-            user_name.setError("Name Empty");
-        }
-        else if(email.isEmpty())
+
+        if(email.isEmpty())
         {
             user_email.setError("Email Empty");
         }
         else if(password.isEmpty())
         {
             user_password.setError("Password Empty");
-        }
-        else if(mobile.isEmpty())
-        {
-            user_mobile.setError("Mobile Empty");
         }
         else if(confirmPwd.isEmpty())
         {
@@ -119,11 +109,11 @@ public class RegisterActivity extends AppCompatActivity {
     private void sendUserData(){
         firebaseDatabase = FirebaseDatabase.getInstance();
         reference = firebaseDatabase.getReference("Chat_Activity").child(firebaseAuth.getUid());
-        reference1 = firebaseDatabase.getReference("Profile_Activity").child(mobile);
+      //  reference1 = firebaseDatabase.getReference("Profile_Activity").child(mobile);
         String token = firebaseAuth.getUid();
-        UserProfile userProfile = new UserProfile(email, username, password, mobile, token);
-        reference.setValue(userProfile);
-        reference1.setValue(userProfile);
+        UserRegistration userRegistration = new UserRegistration(email, password, token);
+        reference.setValue(userRegistration);
+       // reference1.setValue(userProfile);
     }
 
     private void checkConnection() {
